requirements.txt
openai-whisper
torch
torchaudio
fastapi
uvicorn
python-multipart
python-dotenv
numpy
requests
aiofiles
librosa
scipy